Changed name of InventoryScheme.java, Removed and Changed Method to be more modular. Reverse back change made in the FastInv class that are no longer necessary.
